Refactor identity service
